ui: extract display utilities
