+++ src/ui/mod.rs
@@ -1,2 +1,3 @@
+pub mod display;
 pub mod vram_viewer;
 pub mod window;
